requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
dateparser==1.0.0
decorator==4.4.2
fastapi==0.63.0
h11==0.12.0
humanize==3.2.0
idna==2.10
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
log4mongo==1.7.0
maya==0.6.1
multidict==5.1.0
numpy==1.20.0
pandas==1.2.1
parso==0.8.1
pendulum==2.1.2
pickleshare==0.7.5
prompt-toolkit==3.0.16
pycron==3.0.0
pydantic==1.7.3
Pygments==2.8.1
pymongo==3.11.2
python-dateutil==2.8.1
pytz==2021.1
pytzdata==2020.1
pywin32==300
pyzmq==22.0.3
regex==2020.11.13
requests==2.25.1
six==1.15.0
snaptime==0.2.4
starlette==0.13.6
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.26.3
uvicorn==0.13.3
wcwidth==0.2.5
yarl==1.6.3
